@@ -3,105 +3,123 @@
   "version": "0.0.0",
   "main": "server/app.js",
   "dependencies": {
-    "babel-runtime": "^5.8.20",
-    "body-parser": "^1.13.3",
-    "composable-middleware": "^0.3.0",
-    "compression": "^1.5.2",
-    "cookie-parser": "^1.3.5",
-    "errorhandler": "^1.4.2",
     "express": "^4.13.3",
-    "express-jwt": "^3.0.0",
-    "express-sequelize-session": "0.4.0",
-    "express-session": "^1.11.3",
-    "jade": "^1.11.0",
-    "jsonwebtoken": "^5.0.0",
-    "lodash": "^3.10.1",
-    "lusca": "^1.3.0",
+    "morgan": "~1.7.0",
+    "body-parser": "^1.13.3",
     "method-override": "^2.3.5",
-    "morgan": "~1.6.1",
-    "multer": "^1.1.0",
+    "cookie-parser": "^1.3.5",
+    "express-session": "^1.11.3",
+    "errorhandler": "^1.4.2",
+    "compression": "^1.5.2",
+    "composable-middleware": "^0.3.0",
+    "lodash": "^4.6.1",
+    "lusca": "^1.3.0",
+    "babel-runtime": "^6.6.1",
+    "babel-polyfill": "^6.7.2",
+    "jade": "^1.11.0",
+    "sequelize": "^3.5.1",
+    "sqlite3": "~3.1.1",
+    "express-sequelize-session": "0.4.0",
+    "jsonwebtoken": "^7.0.0",
+    "express-jwt": "^3.0.0",
     "passport": "~0.3.0",
     "passport-local": "^1.0.0",
-    "sequelize": "^3.5.1",
-    "serve-favicon": "^2.3.0",
     "socket.io": "^1.3.5",
     "socket.io-client": "^1.3.5",
     "socketio-jwt": "^4.2.0",
-    "sqlite3": "~3.0.2"
+    "serve-favicon": "^2.3.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.0.0",
-    "babel-core": "^5.6.4",
-    "grunt": "~0.4.5",
-    "grunt-wiredep": "^2.0.0",
-    "grunt-concurrent": "^2.0.1",
-    "grunt-contrib-clean": "^0.6.0",
-    "grunt-contrib-concat": "^0.5.1",
-    "grunt-contrib-copy": "^0.8.0",
-    "grunt-contrib-cssmin": "^0.13.0",
-    "grunt-contrib-imagemin": "^0.9.4",
-    "grunt-contrib-jshint": "~0.11.2",
-    "grunt-contrib-uglify": "^0.9.1",
-    "grunt-contrib-watch": "~0.6.1",
-    "grunt-contrib-jade": "^0.15.0",
-    "grunt-babel": "~5.0.0",
-    "grunt-google-cdn": "~0.4.0",
-    "grunt-jscs": "^2.0.0",
-    "grunt-newer": "^1.1.1",
-    "grunt-ng-annotate": "^1.0.1",
-    "grunt-filerev": "^2.3.1",
-    "grunt-usemin": "^3.0.0",
-    "grunt-env": "~0.4.1",
-    "grunt-node-inspector": "^0.2.0",
-    "grunt-nodemon": "^0.4.0",
-    "grunt-angular-templates": "^0.5.4",
-    "grunt-dom-munger": "^3.4.0",
-    "grunt-protractor-runner": "^2.0.0",
-    "grunt-injector": "^0.6.0",
-    "grunt-karma": "~0.12.0",
-    "grunt-build-control": "^0.6.0",
-    "grunt-contrib-stylus": "latest",
-    "jit-grunt": "^0.9.1",
-    "time-grunt": "^1.2.1",
-    "grunt-express-server": "^0.5.1",
-    "grunt-postcss": "^0.6.0",
-    "grunt-open": "~0.2.3",
+    "babel-core": "^6.6.5",
+    "babel-register": "^6.6.5",
+    "babel-plugin-transform-class-properties": "^6.6.0",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-preset-es2015": "^6.6.0",
+    "del": "^2.0.2",
+    "gulp": "^3.9.1",
+    "gulp-add-src": "^0.2.0",
+    "gulp-angular-templatecache": "^1.7.0",
+    "gulp-autoprefixer": "^3.1.0",
+    "gulp-babel": "^6.1.2",
+    "gulp-cache": "^0.4.2",
+    "gulp-concat": "^2.6.0",
+    "gulp-env": "^0.4.0",
+    "gulp-filter": "^4.0.0",
+    "gulp-imagemin": "^2.2.1",
+    "gulp-inject": "^4.0.0",
+    "gulp-istanbul": "~0.10.3",
+    "gulp-istanbul-enforcer": "^1.0.3",
+    "gulp-jscs": "^3.0.2",
+    "gulp-jshint": "^2.0.0",
+    "jshint": "2.9.2",
+    "gulp-livereload": "^3.8.0",
+    "gulp-load-plugins": "^1.0.0-rc.1",
+    "gulp-clean-css": "^2.0.6",
+    "gulp-mocha": "^2.1.3",
+    "gulp-ng-annotate": "^2.0.0",
+    "gulp-ng-constant": "^1.1.0",
+    "gulp-node-inspector": "^0.1.0",
+    "gulp-plumber": "^1.0.1",
+    "gulp-protractor": "^2.1.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-rev": "^7.0.0",
+    "gulp-rev-replace": "^0.4.2",
+    "gulp-sort": "^2.0.0",
+    "gulp-sourcemaps": "^1.5.2",
+    "gulp-svgmin": "^1.1.2",
+    "gulp-uglify": "^1.2.0",
+    "gulp-useref": "^3.0.3",
+    "gulp-util": "^3.0.5",
+    "gulp-watch": "^4.3.5",
+    "gulp-jade": "^1.0.1",
+    "gulp-stylus": "^2.0.4",
+    "gulp-stylint": "^3.0.0",
+    "nib": "^1.1.0",
+    "grunt": "^1.0.1",
+    "grunt-build-control": "^0.7.0",
+    "isparta": "^4.0.0",
+    "utile": "~0.3.0",
+    "nodemon": "^1.3.7",
+    "run-sequence": "^1.1.0",
+    "lazypipe": "^1.0.1",
+    "wiredep": "^2.2.2",
     "open": "~0.0.4",
-    "jshint-stylish": "~2.0.1",
+    "jshint-stylish": "^2.2.0",
     "connect-livereload": "^0.5.3",
-    "mocha": "^2.2.5",
-    "grunt-mocha-test": "~0.12.7",
-    "grunt-mocha-istanbul": "^3.0.1",
-    "istanbul": "^0.3.17",
+    "istanbul": "~0.4.1",
+    "chai": "^3.2.0",
+    "sinon": "^1.16.1",
     "chai-as-promised": "^5.1.0",
     "chai-things": "^0.2.0",
-    "sinon-chai": "^2.8.0",
-    "jasmine-core": "^2.3.4",
-    "karma-jasmine": "~0.3.0",
-    "jasmine-spec-reporter": "^2.4.0",
-    "karma-ng-scenario": "~0.1.0",
-    "karma-firefox-launcher": "~0.1.6",
-    "karma-script-launcher": "~0.1.0",
-    "karma-html2js-preprocessor": "~0.1.0",
-    "karma-ng-jade2js-preprocessor": "^0.2.0",
-    "karma-chrome-launcher": "~0.2.0",
-    "requirejs": "~2.1.11",
-    "karma-requirejs": "~0.2.2",
-    "karma-jade-preprocessor": "0.0.11",
-    "karma-phantomjs-launcher": "~0.2.0",
     "karma": "~0.13.3",
-    "karma-ng-html2js-preprocessor": "~0.1.2",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-script-launcher": "^1.0.0",
+    "karma-chrome-launcher": "^1.0.1",
+    "karma-jade-preprocessor": "0.0.11",
+    "karma-phantomjs-launcher": "~1.0.0",
+    "karma-ng-jade2js-preprocessor": "^0.2.0",
+    "karma-ng-html2js-preprocessor": "^1.0.0",
     "karma-spec-reporter": "~0.0.20",
+    "sinon-chai": "^2.8.0",
+    "mocha": "^2.2.5",
+    "karma-mocha": "^1.0.1",
+    "karma-chai-plugins": "~0.7.0",
+    "karma-babel-preprocessor": "^6.0.1",
+    "phantomjs-prebuilt": "^2.1.4",
     "proxyquire": "^1.0.1",
     "supertest": "^1.1.0"
   },
   "engines": {
-    "node": ">=0.12.0"
+    "node": "^4.4.0",
+    "npm": "^2.14.20"
   },
   "scripts": {
-    "start": "node server",
-    "test": "grunt test",
-    "update-webdriver": "node node_modules/grunt-protractor-runner/node_modules/protractor/bin/webdriver-manager update"
+    "test": "gulp test",
+
+    "update-webdriver": "node node_modules/protractor/bin/webdriver-manager update",
+
+    "start": "node server"
   },
   "private": true
 }
