@@ -9,22 +9,34 @@
 
 'use strict';
 
-var _ = require('lodash');
-var sqldb = require('../../sqldb');
-var Thing = sqldb.Thing;
+import _ from 'lodash';
+import {Thing} from '../../sqldb';
 
-function handleError(res, statusCode) {
-  statusCode = statusCode || 500;
-  return function(err) {
-    res.status(statusCode).send(err);
-  };
-}
-
-function responseWithResult(res, statusCode) {
+function respondWithResult(res, statusCode) {
   statusCode = statusCode || 200;
   return function(entity) {
     if (entity) {
       res.status(statusCode).json(entity);
+    }
+  };
+}
+
+function saveUpdates(updates) {
+  return function(entity) {
+    return entity.updateAttributes(updates)
+      .then(updated => {
+        return updated;
+      });
+  };
+}
+
+function removeEntity(res) {
+  return function(entity) {
+    if (entity) {
+      return entity.destroy()
+        .then(() => {
+          res.status(204).end();
+        });
     }
   };
 }
@@ -39,71 +51,58 @@ function handleEntityNotFound(res) {
   };
 }
 
-function saveUpdates(updates) {
-  return function(entity) {
-    return entity.updateAttributes(updates)
-      .then(function(updated) {
-        return updated;
-      });
-  };
-}
-
-function removeEntity(res) {
-  return function(entity) {
-    if (entity) {
-      return entity.destroy()
-        .then(function() {
-          res.status(204).end();
-        });
-    }
+function handleError(res, statusCode) {
+  statusCode = statusCode || 500;
+  return function(err) {
+    res.status(statusCode).send(err);
   };
 }
 
 // Gets a list of Things
-exports.index = function(req, res) {
-  Thing.findAll()
-    .then(responseWithResult(res))
+export function index(req, res) {
+  return Thing.findAll()
+    .then(respondWithResult(res))
     .catch(handleError(res));
-};
+}
 
 // Gets a single Thing from the DB
-exports.show = function(req, res) {
-  Thing.find({
+export function show(req, res) {
+  return Thing.find({
     where: {
       _id: req.params.id
     }
   })
     .then(handleEntityNotFound(res))
-    .then(responseWithResult(res))
+    .then(respondWithResult(res))
     .catch(handleError(res));
-};
+}
 
 // Creates a new Thing in the DB
-exports.create = function(req, res) {
-  Thing.create(req.body)
-    .then(responseWithResult(res, 201))
+export function create(req, res) {
+  return Thing.create(req.body)
+    .then(respondWithResult(res, 201))
     .catch(handleError(res));
-};
+}
 
 // Updates an existing Thing in the DB
-exports.update = function(req, res) {
+export function update(req, res) {
   if (req.body._id) {
     delete req.body._id;
   }
-  Thing.find({
+  return Thing.find({
     where: {
       _id: req.params.id
     }
   })
     .then(handleEntityNotFound(res))
     .then(saveUpdates(req.body))
-    .then(responseWithResult(res))
+    .then(respondWithResult(res))
     .catch(handleError(res));
-};
+}
 
 // Deletes a Thing from the DB
-exports.destroy = function(req, res) {
-  Thing.find({
+export function destroy(req, res) {
+  return Thing.find({
     where: {
       _id: req.params.id
     }
@@ -111,4 +110,4 @@ exports.destroy = function(req, res) {
     .then(handleEntityNotFound(res))
     .then(removeEntity(res))
     .catch(handleError(res));
-};
+}
