@@ -4,7 +4,7 @@
 
 'use strict';
 
-var EventEmitter = require('events').EventEmitter;
+import {EventEmitter} from 'events';
 var Thing = require('../../sqldb').Thing;
 var ThingEvents = new EventEmitter();
 
@@ -32,4 +32,4 @@ function emitEvent(event) {
   }
 }
 
-module.exports = ThingEvents;
+export default ThingEvents;
