@@ -1,7 +1,7 @@
 'use strict';
 
 var app = require('../..');
-var request = require('supertest');
+import request from 'supertest';
 
 var newThing;
 
@@ -15,7 +15,7 @@ describe('Thing API:', function() {
         .get('/api/things')
         .expect(200)
         .expect('Content-Type', /json/)
-        .end(function(err, res) {
+        .end((err, res) => {
           if (err) {
             return done(err);
           }
@@ -25,7 +25,7 @@ describe('Thing API:', function() {
     });
 
     it('should respond with JSON array', function() {
-      things.should.be.instanceOf(Array);
+      expect(things).to.be.instanceOf(Array);
     });
 
   });
@@ -40,7 +40,7 @@ describe('Thing API:', function() {
         })
         .expect(201)
         .expect('Content-Type', /json/)
-        .end(function(err, res) {
+        .end((err, res) => {
           if (err) {
             return done(err);
           }
@@ -50,8 +50,8 @@ describe('Thing API:', function() {
     });
 
     it('should respond with the newly created thing', function() {
-      newThing.name.should.equal('New Thing');
-      newThing.info.should.equal('This is the brand new thing!!!');
+      expect(newThing.name).to.equal('New Thing');
+      expect(newThing.info).to.equal('This is the brand new thing!!!');
     });
 
   });
@@ -64,7 +64,7 @@ describe('Thing API:', function() {
         .get('/api/things/' + newThing._id)
         .expect(200)
         .expect('Content-Type', /json/)
-        .end(function(err, res) {
+        .end((err, res) => {
           if (err) {
             return done(err);
           }
@@ -78,14 +78,14 @@ describe('Thing API:', function() {
     });
 
     it('should respond with the requested thing', function() {
-      thing.name.should.equal('New Thing');
-      thing.info.should.equal('This is the brand new thing!!!');
+      expect(thing.name).to.equal('New Thing');
+      expect(thing.info).to.equal('This is the brand new thing!!!');
     });
 
   });
 
   describe('PUT /api/things/:id', function() {
-    var updatedThing
+    var updatedThing;
 
     beforeEach(function(done) {
       request(app)
@@ -110,8 +110,8 @@ describe('Thing API:', function() {
     });
 
     it('should respond with the updated thing', function() {
-      updatedThing.name.should.equal('Updated Thing');
-      updatedThing.info.should.equal('This is the updated thing!!!');
+      expect(updatedThing.name).to.equal('Updated Thing');
+      expect(updatedThing.info).to.equal('This is the updated thing!!!');
     });
 
   });
@@ -122,7 +122,7 @@ describe('Thing API:', function() {
       request(app)
         .delete('/api/things/' + newThing._id)
         .expect(204)
-        .end(function(err, res) {
+        .end((err, res) => {
           if (err) {
             return done(err);
           }
@@ -134,7 +134,7 @@ describe('Thing API:', function() {
       request(app)
         .delete('/api/things/' + newThing._id)
         .expect(404)
-        .end(function(err, res) {
+        .end((err, res) => {
           if (err) {
             return done(err);
           }
