@@ -2,7 +2,7 @@
 
 import express from 'express';
 import passport from 'passport';
-import auth from '../auth.service';
+import {signToken} from '../auth.service';
 
 var router = express.Router();
 
@@ -16,9 +16,9 @@ router.post('/', function(req, res, next) {
       return res.status(404).json({message: 'Something went wrong, please try again.'});
     }
 
-    var token = auth.signToken(user._id, user.role);
-    res.json({ token: token });
+    var token = signToken(user._id, user.role);
+    res.json({ token });
   })(req, res, next)
 });
 
-module.exports = router;
+export default router;
