@@ -21,8 +21,20 @@ import sqldb from '../sqldb';
 import expressSequelizeSession from 'express-sequelize-session';
 var Store = expressSequelizeSession(session.Store);
 
-module.exports = function(app) {
+export default function(app) {
   var env = app.get('env');
+
+  if (env === 'development' || env === 'test') {
+    app.use(express.static(path.join(config.root, '.tmp')));
+  }
+
+  if (env === 'production') {
+    app.use(favicon(path.join(config.root, 'client', 'favicon.ico')));
+  }
+
+  app.set('appPath', path.join(config.root, 'client'));
+  app.use(express.static(app.get('appPath')));
+  app.use(morgan('dev'));
 
   app.set('views', config.root + '/server/views');
   app.set('view engine', 'jade');
@@ -33,7 +45,7 @@ module.exports = function(app) {
   app.use(cookieParser());
   app.use(passport.initialize());
 
-  // Persist sessions with mongoStore / sequelizeStore
+  // Persist sessions with MongoStore / sequelizeStore
   // We need to enable sessions for passport-twitter because it's an
   // oauth 1.0 strategy, and Lusca depends on sessions
   app.use(session({
@@ -47,7 +59,7 @@ module.exports = function(app) {
    * Lusca - express server security
    * https://github.com/krakenjs/lusca
    */
-  if ('test' !== env) {
+  if (env !== 'test' && !process.env.SAUCE_USERNAME) {
     app.use(lusca({
       csrf: {
         angular: true
@@ -62,22 +74,17 @@ module.exports = function(app) {
     }));
   }
 
-  app.set('appPath', path.join(config.root, 'client'));
-
-  if ('production' === env) {
-    app.use(favicon(path.join(config.root, 'client', 'favicon.ico')));
-    app.use(express.static(app.get('appPath')));
-    app.use(morgan('dev'));
-  }
-
   if ('development' === env) {
-    app.use(require('connect-livereload')());
+    app.use(require('connect-livereload')({
+      ignore: [
+        /^\/api\/(.*)/,
+        /\.js(\?.*)?$/, /\.css(\?.*)?$/, /\.svg(\?.*)?$/, /\.ico(\?.*)?$/, /\.woff(\?.*)?$/,
+        /\.png(\?.*)?$/, /\.jpg(\?.*)?$/, /\.jpeg(\?.*)?$/, /\.gif(\?.*)?$/, /\.pdf(\?.*)?$/
+      ]
+    }));
   }
 
   if ('development' === env || 'test' === env) {
-    app.use(express.static(path.join(config.root, '.tmp')));
-    app.use(express.static(app.get('appPath')));
-    app.use(morgan('dev'));
     app.use(errorHandler()); // Error handler - has to be last
   }
-};
+}
