@@ -1,7 +1,6 @@
 /**
  * Socket.io configuration
  */
-
 'use strict';
 
 import config from './environment';
@@ -13,17 +12,16 @@ function onDisconnect(socket) {
 // When the user connects.. perform this
 function onConnect(socket) {
   // When the client emits 'info', this listens and executes
-  socket.on('info', function(data) {
+  socket.on('info', data => {
     socket.log(JSON.stringify(data, null, 2));
   });
 
   // Insert sockets below
-  require('../api/file/file.socket').register(socket);
   require('../api/thing/thing.socket').register(socket);
 
 }
 
-module.exports = function(socketio) {
+export default function(socketio) {
   // socket.io (v1.x.x) is powered by debug.
   // In order to see all the debug output, set DEBUG (in server/config/local.env.js) to including the desired scope.
   //
@@ -49,10 +47,8 @@ module.exports = function(socketio) {
       console.log(`SocketIO ${socket.nsp.name} [${socket.address}]`, ...data);
     };
 
-    socket.join
-
     // Call onDisconnect.
-    socket.on('disconnect', function() {
+    socket.on('disconnect', () => {
       onDisconnect(socket);
       socket.log('DISCONNECTED');
     });
@@ -61,4 +57,4 @@ module.exports = function(socketio) {
     onConnect(socket);
     socket.log('CONNECTED');
   });
-};
+}
