@@ -1,13 +1,19 @@
 'use strict';
 
-angular.module('plataformaApp')
-  .controller('AdminCtrl', function($scope, $http, Auth, User) {
+(function() {
 
-    // Use the User $resource to fetch all users
-    $scope.users = User.query();
+  class AdminController {
+    constructor(User) {
+      // Use the User $resource to fetch all users
+      this.users = User.query();
+    }
 
-    $scope.delete = function(user) {
-      User.remove({ id: user._id });
-      $scope.users.splice(this.$index, 1);
-    };
-  });
+    delete(user) {
+      user.$remove();
+      this.users.splice(this.users.indexOf(user), 1);
+    }
+  }
+
+  angular.module('plataformaApp.admin')
+    .controller('AdminController', AdminController);
+})();
