@@ -1,6 +1,6 @@
 'use strict';
 
-describe('Controller: MainController', function() {
+describe('Component: mainComponent', function() {
 
   // load the controller's module
   beforeEach(module('plataformaApp'));
@@ -8,25 +8,30 @@ describe('Controller: MainController', function() {
   beforeEach(module('socketMock'));
 
   var scope;
-  var MainController;
+  var mainComponent;
   var state;
   var $httpBackend;
 
   // Initialize the controller and a mock scope
-  beforeEach(inject(function(_$httpBackend_, $controller, $rootScope, $state) {
+  beforeEach(inject(function(_$httpBackend_, $http, $componentController, $rootScope, $state,
+    socket) {
     $httpBackend = _$httpBackend_;
     $httpBackend.expectGET('/api/things')
       .respond(['HTML5 Boilerplate', 'AngularJS', 'Karma', 'Express']);
 
     scope = $rootScope.$new();
     state = $state;
-    MainController = $controller('MainController', {
-      $scope: scope
+    mainComponent = $componentController('main', {
+      $http: $http,
+      $scope: scope,
+      socket: socket
     });
   }));
 
   it('should attach a list of things to the controller', function() {
+    mainComponent.$onInit();
     $httpBackend.flush();
-    expect(MainController.awesomeThings.length).toBe(4);
+    expect(mainComponent.awesomeThings.length)
+      .to.equal(4);
   });
 });
