@@ -2,11 +2,8 @@
 
 angular.module('plataformaApp')
   .config(function($stateProvider) {
-    $stateProvider
-      .state('main', {
-        url: '/',
-        templateUrl: 'app/main/main.html',
-        controller: 'MainController',
-        controllerAs: 'main'
-      });
+    $stateProvider.state('main', {
+      url: '/',
+      template: '<main></main>'
+    });
   });
