@@ -1,24 +1,17 @@
 'use strict';
 
-angular.module('plataformaApp')
-  .factory('Auth', function Auth($http, User, $cookies, $q) {
-    /**
-     * Return a callback or noop function
-     *
-     * @param  {Function|*} cb - a 'potential' function
-     * @return {Function}
-     */
-    var safeCb = function(cb) {
-      return (angular.isFunction(cb)) ? cb : angular.noop;
-    },
+(function() {
 
-    currentUser = {};
+  function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
+    var safeCb = Util.safeCb;
+    var currentUser = {};
+    var userRoles = appConfig.userRoles || [];
 
-    if ($cookies.get('token')) {
+    if ($cookies.get('token') && $location.path() !== '/logout') {
       currentUser = User.get();
     }
 
-    return {
+    var Auth = {
 
       /**
        * Authenticate user and save token
@@ -27,31 +20,34 @@ angular.module('plataformaApp')
        * @param  {Function} callback - optional, function(error, user)
        * @return {Promise}
        */
-      login: function(user, callback) {
+      login({
+        email,
+        password
+      }, callback) {
         return $http.post('/auth/local', {
-          email: user.email,
-          password: user.password
-        })
-        .then(function(res) {
-          $cookies.put('token', res.data.token);
-          currentUser = User.get();
-          return currentUser.$promise;
-        })
-        .then(function(user) {
-          safeCb(callback)(null, user);
-          return user;
-        })
-        .catch(function(err) {
-          this.logout();
-          safeCb(callback)(err.data);
-          return $q.reject(err.data);
-        }.bind(this));
+            email: email,
+            password: password
+          })
+          .then(res => {
+            $cookies.put('token', res.data.token);
+            currentUser = User.get();
+            return currentUser.$promise;
+          })
+          .then(user => {
+            safeCb(callback)(null, user);
+            return user;
+          })
+          .catch(err => {
+            Auth.logout();
+            safeCb(callback)(err.data);
+            return $q.reject(err.data);
+          });
       },
 
       /**
        * Delete access token and user info
        */
-      logout: function() {
+      logout() {
         $cookies.remove('token');
         currentUser = {};
       },
@@ -63,17 +59,16 @@ angular.module('plataformaApp')
        * @param  {Function} callback - optional, function(error, user)
        * @return {Promise}
        */
-      createUser: function(user, callback) {
-        return User.save(user,
-          function(data) {
+      createUser(user, callback) {
+        return User.save(user, function(data) {
             $cookies.put('token', data.token);
             currentUser = User.get();
             return safeCb(callback)(null, user);
-          },
-          function(err) {
-            this.logout();
+          }, function(err) {
+            Auth.logout();
             return safeCb(callback)(err);
-          }.bind(this)).$promise;
+          })
+          .$promise;
       },
 
       /**
@@ -84,15 +79,18 @@ angular.module('plataformaApp')
        * @param  {Function} callback    - optional, function(error, user)
        * @return {Promise}
        */
-      changePassword: function(oldPassword, newPassword, callback) {
-        return User.changePassword({ id: currentUser._id }, {
-          oldPassword: oldPassword,
-          newPassword: newPassword
-        }, function() {
-          return safeCb(callback)(null);
-        }, function(err) {
-          return safeCb(callback)(err);
-        }).$promise;
+      changePassword(oldPassword, newPassword, callback) {
+        return User.changePassword({
+            id: currentUser._id
+          }, {
+            oldPassword: oldPassword,
+            newPassword: newPassword
+          }, function() {
+            return safeCb(callback)(null);
+          }, function(err) {
+            return safeCb(callback)(err);
+          })
+          .$promise;
       },
 
       /**
@@ -102,17 +100,17 @@ angular.module('plataformaApp')
        * @param  {Function|*} callback - optional, funciton(user)
        * @return {Object|Promise}
        */
-      getCurrentUser: function(callback) {
+      getCurrentUser(callback) {
         if (arguments.length === 0) {
           return currentUser;
         }
 
-        var value = (currentUser.hasOwnProperty('$promise')) ? currentUser.$promise : currentUser;
+        var value = currentUser.hasOwnProperty('$promise') ? currentUser.$promise : currentUser;
         return $q.when(value)
-          .then(function(user) {
+          .then(user => {
             safeCb(callback)(user);
             return user;
-          }, function() {
+          }, () => {
             safeCb(callback)({});
             return {};
           });
@@ -125,37 +123,53 @@ angular.module('plataformaApp')
        * @param  {Function|*} callback - optional, function(is)
        * @return {Bool|Promise}
        */
-      isLoggedIn: function(callback) {
+      isLoggedIn(callback) {
         if (arguments.length === 0) {
           return currentUser.hasOwnProperty('role');
         }
 
-        return this.getCurrentUser(null)
-          .then(function(user) {
+        return Auth.getCurrentUser(null)
+          .then(user => {
             var is = user.hasOwnProperty('role');
             safeCb(callback)(is);
             return is;
           });
       },
 
-       /**
-        * Check if a user is an admin
-        *   (synchronous|asynchronous)
-        *
-        * @param  {Function|*} callback - optional, function(is)
-        * @return {Bool|Promise}
-        */
-      isAdmin: function(callback) {
-        if (arguments.length === 0) {
-          return currentUser.role === 'admin';
+      /**
+       * Check if a user has a specified role or higher
+       *   (synchronous|asynchronous)
+       *
+       * @param  {String}     role     - the role to check against
+       * @param  {Function|*} callback - optional, function(has)
+       * @return {Bool|Promise}
+       */
+      hasRole(role, callback) {
+        var hasRole = function(r, h) {
+          return userRoles.indexOf(r) >= userRoles.indexOf(h);
+        };
+
+        if (arguments.length < 2) {
+          return hasRole(currentUser.role, role);
         }
 
-        return this.getCurrentUser(null)
-          .then(function(user) {
-            var is = user.role === 'admin';
-            safeCb(callback)(is);
-            return is;
+        return Auth.getCurrentUser(null)
+          .then(user => {
+            var has = user.hasOwnProperty('role') ? hasRole(user.role, role) : false;
+            safeCb(callback)(has);
+            return has;
           });
+      },
+
+      /**
+       * Check if a user is an admin
+       *   (synchronous|asynchronous)
+       *
+       * @param  {Function|*} callback - optional, function(is)
+       * @return {Bool|Promise}
+       */
+      isAdmin() {
+        return Auth.hasRole.apply(Auth, [].concat.apply(['admin'], arguments));
       },
 
       /**
@@ -163,8 +177,14 @@ angular.module('plataformaApp')
        *
        * @return {String} - a token string used for authenticating
        */
-      getToken: function() {
+      getToken() {
         return $cookies.get('token');
       }
     };
-  });
+
+    return Auth;
+  }
+
+  angular.module('plataformaApp.auth')
+    .factory('Auth', AuthService);
+})();
