@@ -1,10 +1,9 @@
 'use strict';
 
 angular.module('plataformaApp')
-  .directive('navbar', function () {
-    return {
-      templateUrl: 'components/navbar/navbar.html',
-      restrict: 'E',
-      controller: 'NavbarCtrl'
-    };
-  });
+  .directive('navbar', () => ({
+    templateUrl: 'components/navbar/navbar.html',
+    restrict: 'E',
+    controller: 'NavbarController',
+    controllerAs: 'nav'
+  }));
