@@ -7,8 +7,9 @@ describe('Login View', function() {
   var page;
 
   var loadPage = function() {
-    browser.get(config.baseUrl + '/login');
+    let promise = browser.get(config.baseUrl + '/login');
     page = require('./login.po');
+    return promise;
   };
 
   var testUser = {
@@ -17,22 +18,26 @@ describe('Login View', function() {
     password: 'test'
   };
 
-  beforeEach(function(done) {
-    UserModel.destroy({ where: {} })
+  before(function() {
+    return UserModel
+      .destroy({ where: {} })
       .then(function() {
         return UserModel.create(testUser);
       })
-      .then(loadPage)
-      .finally(done);
+      .then(loadPage);
+  });
+
+  after(function() {
+    return UserModel.destroy({ where: {} });
   });
 
   it('should include login form with correct inputs and submit button', function() {
-    expect(page.form.email.getAttribute('type')).toBe('email');
-    expect(page.form.email.getAttribute('name')).toBe('email');
-    expect(page.form.password.getAttribute('type')).toBe('password');
-    expect(page.form.password.getAttribute('name')).toBe('password');
-    expect(page.form.submit.getAttribute('type')).toBe('submit');
-    expect(page.form.submit.getText()).toBe('Login');
+    expect(page.form.email.getAttribute('type')).to.eventually.equal('email');
+    expect(page.form.email.getAttribute('name')).to.eventually.equal('email');
+    expect(page.form.password.getAttribute('type')).to.eventually.equal('password');
+    expect(page.form.password.getAttribute('name')).to.eventually.equal('password');
+    expect(page.form.submit.getAttribute('type')).to.eventually.equal('submit');
+    expect(page.form.submit.getText()).to.eventually.equal('Login');
   });
 
   describe('with local auth', function() {
@@ -42,20 +47,27 @@ describe('Login View', function() {
 
       var navbar = require('../../components/navbar/navbar.po');
 
-      expect(browser.getCurrentUrl()).toBe(config.baseUrl + '/');
-      expect(navbar.navbarAccountGreeting.getText()).toBe('Hello ' + testUser.name);
+      expect(browser.getCurrentUrl()).to.eventually.equal(config.baseUrl + '/');
+      expect(navbar.navbarAccountGreeting.getText()).to.eventually.equal('Hello ' + testUser.name);
     });
 
-    it('should indicate login failures', function() {
-      page.login({
-        email: testUser.email,
-        password: 'badPassword'
+    describe('and invalid credentials', function() {
+      before(function() {
+        return loadPage();
+      })
+
+      it('should indicate login failures', function() {
+        page.login({
+          email: testUser.email,
+          password: 'badPassword'
+        });
+
+        expect(browser.getCurrentUrl()).to.eventually.equal(config.baseUrl + '/login');
+
+        var helpBlock = page.form.element(by.css('.form-group.has-error .help-block.ng-binding'));
+        expect(helpBlock.getText()).to.eventually.equal('This password is not correct.');
       });
 
-      expect(browser.getCurrentUrl()).toBe(config.baseUrl + '/login');
-
-      var helpBlock = page.form.element(by.css('.form-group.has-error .help-block.ng-binding'));
-      expect(helpBlock.getText()).toBe('This password is not correct.');
     });
 
   });
