@@ -7,10 +7,10 @@
 
 var SignupPage = function() {
   var form = this.form = element(by.css('.form'));
-  form.name = form.element(by.model('user.name'));
-  form.email = form.element(by.model('user.email'));
-  form.password = form.element(by.model('user.password'));
-  form.confirmPassword = form.element(by.model('user.confirmPassword'));
+  form.name = form.element(by.model('vm.user.name'));
+  form.email = form.element(by.model('vm.user.email'));
+  form.password = form.element(by.model('vm.user.password'));
+  form.confirmPassword = form.element(by.model('vm.user.confirmPassword'));
   form.submit = form.element(by.css('.btn-register'));
 
   this.signup = function(data) {
@@ -21,7 +21,7 @@ var SignupPage = function() {
       }
     }
 
-    form.submit.click();
+    return form.submit.click();
   };
 };
 
