@@ -7,9 +7,10 @@ describe('Signup View', function() {
   var page;
 
   var loadPage = function() {
-    browser.manage().deleteAllCookies();
-    browser.get(config.baseUrl + '/signup');
+    browser.manage().deleteAllCookies()
+    let promise = browser.get(config.baseUrl + '/signup');
     page = require('./signup.po');
+    return promise;
   };
 
   var testUser = {
@@ -19,46 +20,57 @@ describe('Signup View', function() {
     confirmPassword: 'test'
   };
 
-  beforeEach(function() {
-    loadPage();
+  before(function() {
+    return loadPage();
+  });
+
+  after(function() {
+    return UserModel.destroy({ where: {} });
   });
 
   it('should include signup form with correct inputs and submit button', function() {
-    expect(page.form.name.getAttribute('type')).toBe('text');
-    expect(page.form.name.getAttribute('name')).toBe('name');
-    expect(page.form.email.getAttribute('type')).toBe('email');
-    expect(page.form.email.getAttribute('name')).toBe('email');
-    expect(page.form.password.getAttribute('type')).toBe('password');
-    expect(page.form.password.getAttribute('name')).toBe('password');
-    expect(page.form.confirmPassword.getAttribute('type')).toBe('password');
-    expect(page.form.confirmPassword.getAttribute('name')).toBe('confirmPassword');
-    expect(page.form.submit.getAttribute('type')).toBe('submit');
-    expect(page.form.submit.getText()).toBe('Sign up');
+    expect(page.form.name.getAttribute('type')).to.eventually.equal('text');
+    expect(page.form.name.getAttribute('name')).to.eventually.equal('name');
+    expect(page.form.email.getAttribute('type')).to.eventually.equal('email');
+    expect(page.form.email.getAttribute('name')).to.eventually.equal('email');
+    expect(page.form.password.getAttribute('type')).to.eventually.equal('password');
+    expect(page.form.password.getAttribute('name')).to.eventually.equal('password');
+    expect(page.form.confirmPassword.getAttribute('type')).to.eventually.equal('password');
+    expect(page.form.confirmPassword.getAttribute('name')).to.eventually.equal('confirmPassword');
+    expect(page.form.submit.getAttribute('type')).to.eventually.equal('submit');
+    expect(page.form.submit.getText()).to.eventually.equal('Sign up');
   });
 
   describe('with local auth', function() {
 
-    beforeAll(function(done) {
-      UserModel.destroy({ where: {} }).then(done);
-    });
+    before(function() {
+      return UserModel.destroy({ where: {} });
+    })
 
     it('should signup a new user, log them in, and redirecting to "/"', function() {
       page.signup(testUser);
 
       var navbar = require('../../components/navbar/navbar.po');
 
-      expect(browser.getCurrentUrl()).toBe(config.baseUrl + '/');
-      expect(navbar.navbarAccountGreeting.getText()).toBe('Hello ' + testUser.name);
+      expect(browser.getCurrentUrl()).to.eventually.equal(config.baseUrl + '/');
+      expect(navbar.navbarAccountGreeting.getText()).to.eventually.equal('Hello ' + testUser.name);
     });
 
-    it('should indicate signup failures', function() {
-      page.signup(testUser);
+    describe('and invalid credentials', function() {
+      before(function() {
+        return loadPage();
+      });
 
-      expect(browser.getCurrentUrl()).toBe(config.baseUrl + '/signup');
-      expect(page.form.email.getAttribute('class')).toContain('ng-invalid-mongoose');
+      it('should indicate signup failures', function() {
+        page.signup(testUser);
 
-      var helpBlock = page.form.element(by.css('.form-group.has-error .help-block.ng-binding'));
-      expect(helpBlock.getText()).toBe('The specified email address is already in use.');
+        expect(browser.getCurrentUrl()).to.eventually.equal(config.baseUrl + '/signup');
+        expect(page.form.email.getAttribute('class')).to.eventually.contain('ng-invalid-mongoose');
+
+        var helpBlock = page.form.element(by.css('.form-group.has-error .help-block.ng-binding'));
+        expect(helpBlock.getText()).to.eventually.equal('The specified email address is already in use.');
+      });
+
     });
 
   });
