@@ -6,13 +6,14 @@ describe('Main View', function() {
   var page;
 
   beforeEach(function() {
-    browser.get(config.baseUrl + '/');
+    let promise = browser.get(config.baseUrl + '/');
     page = require('./main.po');
+    return promise;
   });
 
   it('should include jumbotron with correct data', function() {
-    expect(page.h1El.getText()).toBe('\'Allo, \'Allo!');
-    expect(page.imgEl.getAttribute('src')).toMatch(/yeoman.png$/);
-    expect(page.imgEl.getAttribute('alt')).toBe('I\'m Yeoman');
+    expect(page.h1El.getText()).to.eventually.equal('\'Allo, \'Allo!');
+    expect(page.imgEl.getAttribute('src')).to.eventually.match(/yeoman(\.[a-zA-Z0-9]*)?\.png$/);
+    expect(page.imgEl.getAttribute('alt')).to.eventually.equal('I\'m Yeoman');
   });
 });
