@@ -12,10 +12,10 @@ var userCtrlStub = {
 };
 
 var authServiceStub = {
-  isAuthenticated: function() {
+  isAuthenticated() {
     return 'authService.isAuthenticated';
   },
-  hasRole: function(role) {
+  hasRole(role) {
     return 'authService.hasRole.' + role;
   }
 };
@@ -30,7 +30,7 @@ var routerStub = {
 // require the index with our stubbed out modules
 var userIndex = proxyquire('./index', {
   'express': {
-    Router: function() {
+    Router() {
       return routerStub;
     }
   },
@@ -41,15 +41,15 @@ var userIndex = proxyquire('./index', {
 describe('User API Router:', function() {
 
   it('should return an express router instance', function() {
-    userIndex.should.equal(routerStub);
+    expect(userIndex).to.equal(routerStub);
   });
 
   describe('GET /api/users', function() {
 
     it('should verify admin role and route to user.controller.index', function() {
-      routerStub.get
+      expect(routerStub.get
         .withArgs('/', 'authService.hasRole.admin', 'userCtrl.index')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
@@ -57,9 +57,9 @@ describe('User API Router:', function() {
   describe('DELETE /api/users/:id', function() {
 
     it('should verify admin role and route to user.controller.destroy', function() {
-      routerStub.delete
+      expect(routerStub.delete
         .withArgs('/:id', 'authService.hasRole.admin', 'userCtrl.destroy')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
@@ -67,9 +67,9 @@ describe('User API Router:', function() {
   describe('GET /api/users/me', function() {
 
     it('should be authenticated and route to user.controller.me', function() {
-      routerStub.get
+      expect(routerStub.get
         .withArgs('/me', 'authService.isAuthenticated', 'userCtrl.me')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
@@ -77,9 +77,9 @@ describe('User API Router:', function() {
   describe('PUT /api/users/:id/password', function() {
 
     it('should be authenticated and route to user.controller.changePassword', function() {
-      routerStub.put
+      expect(routerStub.put
         .withArgs('/:id/password', 'authService.isAuthenticated', 'userCtrl.changePassword')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
@@ -87,9 +87,9 @@ describe('User API Router:', function() {
   describe('GET /api/users/:id', function() {
 
     it('should be authenticated and route to user.controller.show', function() {
-      routerStub.get
+      expect(routerStub.get
         .withArgs('/:id', 'authService.isAuthenticated', 'userCtrl.show')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
@@ -97,9 +97,9 @@ describe('User API Router:', function() {
   describe('POST /api/users', function() {
 
     it('should route to user.controller.create', function() {
-      routerStub.post
+      expect(routerStub.post
         .withArgs('/', 'userCtrl.create')
-        .should.have.been.calledOnce;
+        ).to.have.been.calledOnce;
     });
 
   });
