@@ -19,19 +19,12 @@ function handleError(res, statusCode) {
   };
 }
 
-function respondWith(res, statusCode) {
-  statusCode = statusCode || 200;
-  return function() {
-    res.status(statusCode).end();
-  };
-}
-
 /**
  * Get list of users
  * restriction: 'admin'
  */
-exports.index = function(req, res) {
-  User.findAll({
+export function index(req, res) {
+  return User.findAll({
     attributes: [
       '_id',
       'name',
@@ -40,81 +33,79 @@ exports.index = function(req, res) {
       'provider'
     ]
   })
-    .then(function(users) {
+    .then(users => {
       res.status(200).json(users);
     })
     .catch(handleError(res));
-};
+}
 
 /**
  * Creates a new user
  */
-exports.create = function(req, res, next) {
+export function create(req, res, next) {
   var newUser = User.build(req.body);
   newUser.setDataValue('provider', 'local');
   newUser.setDataValue('role', 'user');
-  newUser.save()
+  return newUser.save()
     .then(function(user) {
       var token = jwt.sign({ _id: user._id }, config.secrets.session, {
-        expiresInMinutes: 60 * 5
+        expiresIn: 60 * 60 * 5
       });
-      res.json({ token: token });
+      res.json({ token });
     })
     .catch(validationError(res));
-};
+}
 
 /**
  * Get a single user
  */
-exports.show = function(req, res, next) {
+export function show(req, res, next) {
   var userId = req.params.id;
 
-  User.find({
+  return User.find({
     where: {
       _id: userId
     }
   })
-    .then(function(user) {
+    .then(user => {
       if (!user) {
         return res.status(404).end();
       }
       res.json(user.profile);
     })
-    .catch(function(err) {
-      return next(err);
-    });
-};
+    .catch(err => next(err));
+}
 
 /**
  * Deletes a user
  * restriction: 'admin'
  */
-exports.destroy = function(req, res) {
-  User.destroy({ _id: req.params.id })
+export function destroy(req, res) {
+  return User.destroy({ _id: req.params.id })
     .then(function() {
       res.status(204).end();
     })
     .catch(handleError(res));
-};
+}
 
 /**
  * Change a users password
  */
-exports.changePassword = function(req, res, next) {
+export function changePassword(req, res, next) {
   var userId = req.user._id;
   var oldPass = String(req.body.oldPassword);
   var newPass = String(req.body.newPassword);
 
-  User.find({
+  return User.find({
     where: {
       _id: userId
     }
   })
-    .then(function(user) {
+    .then(user => {
       if (user.authenticate(oldPass)) {
         user.password = newPass;
         return user.save()
-          .then(function() {
+          .then(() => {
             res.status(204).end();
           })
           .catch(validationError(res));
@@ -122,15 +113,15 @@ exports.changePassword = function(req, res, next) {
         return res.status(403).end();
       }
     });
-};
+}
 
 /**
  * Get my info
  */
-exports.me = function(req, res, next) {
+export function me(req, res, next) {
   var userId = req.user._id;
 
-  User.find({
+  return User.find({
     where: {
       _id: userId
     },
@@ -142,20 +133,18 @@ exports.me = function(req, res, next) {
       'provider'
     ]
   })
-    .then(function(user) { // don't ever give out the password or salt
+    .then(user => { // don't ever give out the password or salt
       if (!user) {
         return res.status(401).end();
       }
       res.json(user);
     })
-    .catch(function(err) {
-      return next(err);
-    });
-};
+    .catch(err => next(err));
+}
 
 /**
  * Authentication callback
  */
-exports.authCallback = function(req, res, next) {
+export function authCallback(req, res, next) {
   res.redirect('/');
-};
+}
