@@ -6,6 +6,7 @@ import jwt from 'jsonwebtoken';
 import expressJwt from 'express-jwt';
 import compose from 'composable-middleware';
 import {User} from'../sqldb';
+
 var validateJwt = expressJwt({
   secret: config.secrets.session
 });
@@ -14,7 +15,7 @@ var validateJwt = expressJwt({
  * Attaches the user object to the request if authenticated
  * Otherwise returns 403
  */
-function isAuthenticated() {
+export function isAuthenticated() {
   return compose()
     // Validate jwt
     .use(function(req, res, next) {
@@ -31,23 +32,21 @@ function isAuthenticated() {
           _id: req.user._id
         }
       })
-        .then(function(user) {
+        .then(user => {
           if (!user) {
             return res.status(401).end();
           }
           req.user = user;
           next();
         })
-        .catch(function(err) {
-          return next(err);
-        });
+        .catch(err => next(err));
     });
 }
 
 /**
  * Checks if the user role meets the minimum requirements of the route
  */
-function hasRole(roleRequired) {
+export function hasRole(roleRequired) {
   if (!roleRequired) {
     throw new Error('Required role needs to be set');
   }
@@ -58,8 +57,7 @@ function hasRole(roleRequired) {
       if (config.userRoles.indexOf(req.user.role) >=
           config.userRoles.indexOf(roleRequired)) {
         next();
-      }
-      else {
+      } else {
         res.status(403).send('Forbidden');
       }
     });
@@ -68,25 +66,20 @@ function hasRole(roleRequired) {
 /**
  * Returns a jwt token signed by the app secret
  */
-function signToken(id, role) {
+export function signToken(id, role) {
   return jwt.sign({ _id: id, role: role }, config.secrets.session, {
-    expiresInMinutes: 60 * 180
+    expiresIn: 60 * 60 * 5
   });
 }
 
 /**
  * Set token cookie directly for oAuth strategies
  */
-function setTokenCookie(req, res) {
+export function setTokenCookie(req, res) {
   if (!req.user) {
-    return res.status(404).send('Something went wrong, please try again.');
+    return res.status(404).send('It looks like you aren\'t logged in, please try again.');
   }
   var token = signToken(req.user._id, req.user.role);
   res.cookie('token', token);
   res.redirect('/');
 }
-
-exports.isAuthenticated = isAuthenticated;
-exports.hasRole = hasRole;
-exports.signToken = signToken;
-exports.setTokenCookie = setTokenCookie;
