@@ -7,7 +7,7 @@ function localAuthenticate(User, email, password, done) {
       email: email.toLowerCase()
     }
   })
-    .then(function(user) {
+    .then(user => {
       if (!user) {
         return done(null, false, {
           message: 'This email is not registered.'
@@ -18,24 +18,20 @@ function localAuthenticate(User, email, password, done) {
           return done(authError);
         }
         if (!authenticated) {
-          return done(null, false, {
-            message: 'This password is not correct.'
-          });
+          return done(null, false, { message: 'This password is not correct.' });
         } else {
           return done(null, user);
         }
       });
     })
-    .catch(function(err) {
-      return done(err);
-    });
+    .catch(err => done(err));
 }
 
-exports.setup = function(User, config) {
+export function setup(User, config) {
   passport.use(new LocalStrategy({
     usernameField: 'email',
     passwordField: 'password' // this is the virtual field on the model
   }, function(email, password, done) {
     return localAuthenticate(User, email, password, done);
   }));
-};
+}
