@@ -14,7 +14,7 @@ module.exports[404] = function pageNotFound(req, res) {
   res.status(result.status);
   res.render(viewFilePath, {}, function(err, html) {
     if (err) {
-      return res.json(result, result.status);
+      return res.status(result.status).json(result);
     }
 
     res.send(html);
