@@ -4,21 +4,24 @@
 // =================================
 module.exports = {
   // Server IP
-  ip:       process.env.OPENSHIFT_NODEJS_IP ||
-            process.env.IP ||
-            undefined,
+  ip:     process.env.OPENSHIFT_NODEJS_IP ||
+          process.env.IP ||
+          undefined,
 
   // Server port
-  port:     process.env.OPENSHIFT_NODEJS_PORT ||
-            process.env.PORT ||
-            8080,
+  port:   process.env.OPENSHIFT_NODEJS_PORT ||
+          process.env.PORT ||
+          8080,
 
-  // MongoDB connection options
-  mongo: {
-    uri:    process.env.MONGOLAB_URI ||
-            process.env.MONGOHQ_URL ||
-            process.env.OPENSHIFT_MONGODB_DB_URL +
-            process.env.OPENSHIFT_APP_NAME ||
-            'mongodb://localhost/plataforma'
+  sequelize: {
+    uri:  process.env.SEQUELIZE_URI ||
+          'sqlite://',
+    options: {
+      logging: false,
+      storage: 'dist.sqlite',
+      define: {
+        timestamps: false
+      }
+    }
   }
 };
