@@ -5,7 +5,7 @@ var env = process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
 if (env === 'development' || env === 'test') {
   // Register the Babel require hook
-  require('babel-core/register');
+  require('babel-register');
 }
 
 // Export the application
