@@ -9,12 +9,11 @@ import config from '../config/environment';
 import Sequelize from 'sequelize';
 
 var db = {
-  Sequelize: Sequelize,
+  Sequelize,
   sequelize: new Sequelize(config.sequelize.uri, config.sequelize.options)
 };
 
 // Insert models below
-db.File = db.sequelize.import('../api/file/file.model');
 db.Thing = db.sequelize.import('../api/thing/thing.model');
 db.User = db.sequelize.import('../api/user/user.model');
 
